requirements.txt
vit-pytorch
torch
torchvision
numpy
pandas
pillow
timm
